# [SYSTEMDS-3610] BuiltinSwitch Aggregate Skip