I have observed depending on JVM the switch case in the builtin objects is taking a long time. we can remove this by having dedicated objects.